+++ ovp_search/views.py
@@ -34,8 +34,15 @@ class OrganizationSearchResource(mixins.ListModelMixin, viewsets.GenericViewSet)
       cause = params.get('cause', None)
       address = params.get('address', None)
       name = params.get('name', None)
+      published = params.get('published', 'true')
 
       queryset = SearchQuerySet().models(Organization)
+
+      if published == "true":
+        queryset = queryset.filter(published=True)
+      elif published == "false":
+        queryset = queryset.filter(published=False)
+      # Any other value will return both published and unpublished
 
       if name:
         queryset.filter(name=name)
@@ -86,7 +93,7 @@ class OrganizationSearchResource(mixins.ListModelMixin, viewsets.GenericViewSet)
       # to work properly with django-rest-framework
       # TODO: Find a solution
       result_keys = [q.pk for q in queryset]
-      result = Organization.objects.filter(pk__in=result_keys, deleted=False, published=True).prefetch_related('causes').select_related('address').order_by('-highlighted')
+      result = Organization.objects.filter(pk__in=result_keys, deleted=False).prefetch_related('causes').select_related('address').order_by('-highlighted')
       cache.set(key, result, cache_ttl)
 
     return result
@@ -112,8 +119,15 @@ class ProjectSearchResource(mixins.ListModelMixin, viewsets.GenericViewSet):
       address = params.get('address', None)
       highlighted = (params.get('highlighted') == 'true')
       name = params.get('name', None)
+      published = params.get('published', 'true')
 
       queryset = SearchQuerySet().models(Project)
+
+      if published == "true":
+        queryset = queryset.filter(published=True)
+      elif published == "false":
+        queryset = queryset.filter(published=False)
+      # Any other value will return both published and unpublished
 
       # TODO: Implement when Organization has slug
       #if nonprofit:
@@ -179,7 +193,7 @@ class ProjectSearchResource(mixins.ListModelMixin, viewsets.GenericViewSet):
       # to work properly with django-rest-framework
       # TODO: Find a solution
       result_keys = [q.pk for q in queryset]
-      result = Project.objects.filter(pk__in=result_keys, deleted=False, published=True, closed=False).prefetch_related('skills', 'causes').select_related('address').order_by('-highlighted')
+      result = Project.objects.filter(pk__in=result_keys, deleted=False, closed=False).prefetch_related('skills', 'causes').select_related('address').order_by('-highlighted')
       cache.set(key, result, cache_ttl)
 
     return result